Remove unused metrics collection resources (#4475)
* Remove unused metrics collection resources

* Strip out extraneous Docker Compose file for metrics
* Remove StatsD configs
* Adjust Vagrantfile and STRTA to remove references to metrics

* Remove unused DIR variable from scripts/server

DIR was previously telling scripts/server how to access Docker Compose
override files. Now that the overrides are no longer necessary, remove
the variable declaration in order to appease shellcheck.

* Update CHANGELOG

Add entry for removing unused metrics collection resources.

* Remove AspectJ installation from Dockerfiles

Since tracing is no longer functional, remove the installation and
configuration of AspectJ from API and tile Dockerfiles.

* Suppress warning that AspectJ is missing in the dev env

Instruct Kamon not to warn the dev user that it was not initialized
with AspectJ.